Moved random acquisition function and DTS to individual files
